+++ src/components/super/i-static-page/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.33.0 (2022-12-28)
+
+#### :rocket: New Feature
+
+* Added the ability to manipulate meta information of a page
+
 ## v3.29.0 (2022-09-13)
 
 #### :boom: Breaking Change